warnings

+++ src/glib/base/dt.h
@@ -502,7 +502,7 @@ public:
   /// Get last character in string (before null terminator)
   char LastCh() const {return GetCh(Len()-1);}
   /// Get String Length (null terminator not included)
-  int Len() const { return Empty() ? 0 : strlen(Inner); }
+  int Len() const { return Empty() ? 0 : (int)strlen(Inner); }
   /// Check if this is an empty string
   bool Empty() const { IAssertR(Inner == nullptr || Inner[0] != 0, "TStr::Empty string is not nullptr. Fix immediately!");  return  Inner == nullptr; }
   /// Deletes the char pointer if it is not nullptr. (not thread safe)
